+++ source/Delphi.Forms.JSON.pas
@@ -6,9 +6,18 @@ uses
   System.JSON,
   Delphi.Forms.Types;
 
-function FormFileToJSON(FormFile: TFormFile): TJSONObject;
-function FormObjectToJSON(Obj: TFormObject): TJSONObject;
-function FormPropertyToJSON(Prop: TFormProperty): TJSONObject;
+type
+
+  TFormJsonOption = (joIncludeValues, joIncludePositions, joPrettyPrint, joIncludeRawText);
+  TFormJsonOptions = set of TFormJsonOption;
+
+const
+  DefaultJsonOptions: TFormJsonOptions = [joIncludeValues, joPrettyPrint];
+
+function FormFileToJSON(FormFile: TFormFile; Options: TFormJsonOptions = [joIncludeValues, joPrettyPrint]): TJSONObject;
+function FormFileToJSONString(FormFile: TFormFile; Options: TFormJsonOptions = [joIncludeValues, joPrettyPrint]): string;
+function FormObjectToJSON(Obj: TFormObject; Options: TFormJsonOptions = [joIncludeValues, joPrettyPrint]): TJSONObject;
+function FormPropertyToJSON(Prop: TFormProperty; Options: TFormJsonOptions = [joIncludeValues, joPrettyPrint]): TJSONObject;
 function FormValuePreview(Value: TFormValue): string;
 function FormValueKindName(Kind: TFormValueKind): string;
 function FormObjectKindName(Kind: TObjectKind): string;
@@ -94,15 +103,24 @@ begin
   end;
 end;
 
-function FormPropertyToJSON(Prop: TFormProperty): TJSONObject;
+function FormPropertyToJSON(Prop: TFormProperty; Options: TFormJsonOptions): TJSONObject;
 begin
   Result := TJSONObject.Create;
   Result.AddPair('name', Prop.Name);
   Result.AddPair('kind', FormValueKindName(Prop.Value.Kind));
-  Result.AddPair('value', FormValuePreview(Prop.Value));
+  if joIncludeValues in Options then
+    Result.AddPair('value', FormValuePreview(Prop.Value));
+  if joIncludeRawText in Options then
+    if Prop.Value.RawText <> '' then
+      Result.AddPair('rawText', Prop.Value.RawText);
+  if joIncludePositions in Options then
+  begin
+    Result.AddPair('sourceStart', TJSONNumber.Create(Prop.SourceStart));
+    Result.AddPair('sourceEnd', TJSONNumber.Create(Prop.SourceEnd));
+  end;
 end;
 
-function FormObjectToJSON(Obj: TFormObject): TJSONObject;
+function FormObjectToJSON(Obj: TFormObject; Options: TFormJsonOptions): TJSONObject;
 var
   Props: TJSONArray;
   Children: TJSONArray;
@@ -112,21 +130,41 @@ begin
   Result.AddPair('objectKind', FormObjectKindName(Obj.ObjectKind));
   Result.AddPair('name', Obj.Name);
   Result.AddPair('className', Obj.ClassName_);
+  if joIncludePositions in Options then
+  begin
+    Result.AddPair('sourceStart', TJSONNumber.Create(Obj.SourceStart));
+    Result.AddPair('sourceEnd', TJSONNumber.Create(Obj.SourceEnd));
+  end;
   Props := TJSONArray.Create;
   for I := 0 to Obj.Properties.Count - 1 do
-    Props.AddElement(FormPropertyToJSON(Obj.Properties[I]));
+    Props.AddElement(FormPropertyToJSON(Obj.Properties[I], Options));
   Result.AddPair('properties', Props);
   Children := TJSONArray.Create;
   for I := 0 to Obj.Children.Count - 1 do
-    Children.AddElement(FormObjectToJSON(Obj.Children[I]));
+    Children.AddElement(FormObjectToJSON(Obj.Children[I], Options));
   Result.AddPair('children', Children);
 end;
 
-function FormFileToJSON(FormFile: TFormFile): TJSONObject;
+function FormFileToJSON(FormFile: TFormFile; Options: TFormJsonOptions): TJSONObject;
 begin
   Result := TJSONObject.Create;
   if FormFile.Root <> nil then
-    Result.AddPair('root', FormObjectToJSON(FormFile.Root));
+    Result.AddPair('root', FormObjectToJSON(FormFile.Root, Options));
+end;
+
+function FormFileToJSONString(FormFile: TFormFile; Options: TFormJsonOptions): string;
+var
+  Json: TJSONObject;
+begin
+  Json := FormFileToJSON(FormFile, Options);
+  try
+    if joPrettyPrint in Options then
+      Result := Json.Format(2)
+    else
+      Result := Json.ToJSON;
+  finally
+    Json.Free;
+  end;
 end;
 
 end.